% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.mass.R
\name{association.matrix}
\alias{association.matrix}
\title{Association Matrix}
\usage{
association.matrix(cl)
}
\arguments{
\item{cl}{A vector containing cluster labels for a clustering/partition.}
}
\value{
A matrix of 0s and 1s indicating whether items i and j are in
the same cluster/subset.
}
\description{
This function creates an association matrix for a clustering/partition. 
The (i,j) element of the matrix is 1 if item i and j are in the same 
cluster/subset and 0 otherwise.
}
\examples{

cl <- rep(1:3,times=c(2,4,3))
association.matrix(cl)

}
