% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shallot.R
\name{enumerate.partitions}
\alias{enumerate.partitions}
\title{Enumerate Partitions}
\usage{
enumerate.partitions(n.items)
}
\arguments{
\item{n.items}{An integer given then number of items to partition.}
}
\value{
A matrix of cluster labels in which each row represents a
clusterings.
}
\description{
This function enumerates all possible partitions for a given number of
items.
}
\details{
This function returns an enumeration of the partition of \code{n.items}
items.
}
\examples{

\dontrun{
example(shallot)
}

}
\seealso{
\code{\link{process.samples}}
}
\author{
David B. Dahl \email{dahl@stat.byu.edu}
}
