% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shallot.R
\name{partition.pmf}
\alias{partition.pmf}
\title{Obtain the Probability Mass Function of a Partition Distribution}
\usage{
partition.pmf(x)
}
\arguments{
\item{x}{An object of class \code{shallot.distribution} obtained, for
example, from the \code{\link{ewens.pitman.attraction}} function.}
}
\value{
A function that takes a partition (as a vector in cluster label
notation) and returns the probability --- or, if \code{log=TRUE}, the log of
the probability --- of the supplied partition.
}
\description{
This function returns the probability mass function (pmf) of a partition
distribution.
}
\examples{

\dontrun{
example(shallot)
}

}
\author{
David B. Dahl \email{dahl@stat.byu.edu}
}
